+++ Changelog.md
@@ -1,6 +1,8 @@
 # 1.1.0
 
 - Update Path Matcher to support AnyAfter segments (**)
+- Remove Test Example
+- Add Paste Bin App Example
 
 
 # 1.0.0!